Refactor: Handle action in ListScreen with LaunchedEffect
The `action` is now passed to the `ListScreen` composable.
`action` is now handled in the `LaunchedEffect` in the `ListScreen`.
`getAllTasks` and `readSortState` are called once during the initialization of the `SharedViewModel`
`handleDatabaseAction` is called when the action changes in `ListScreen`.
`DisplaySnackBar` calls `onComplete` when it finishes.
The `action` value is set to `ToDoAction.NO_ACTION` after the snackbar is displayed.